Fix proposal creation for SimpleProposal x Awesome proposal private fields - remove old proposals_controller.rb overload - overload SimpleProposals proposals_controller_override.rb
